login is now on a different file, main file uses new login page and displays it
